Fix GU_CHAR_ACTION_ATTACK structure

+++ DboShared/NtlShared2/NtlPacketGU.h
@@ -1076,7 +1076,7 @@ BEGIN_PROTOCOL(GU_CHAR_ACTION_ATTACK)
 	BYTE					byAttackResult:7; // Action Result ( eBATTLE_ATTACK_RESULT )
 	int						attackResultValue; // Result of the action (damage) //updated
 	BYTE					byAttackSequence; // Attack sequence number (chain attack)
-	float					fReflectedDamage; // Damage and is reflected from the target.
+	int					    fReflectedDamage; // Damage and is reflected from the target.
 	BYTE					byBlockedAction;		// eDBO_GUARD_TYPE
 	sVECTOR3				vShift;
 	sDBO_LP_EP_RECOVERED	lpEpRecovered; //new
